+++ lib/react_on_rails/server_rendering_pool.rb
@@ -9,12 +9,8 @@ require_relative "server_rendering_pool/ruby_embedded_java_script"
 module ReactOnRails
   module ServerRenderingPool
     class << self
-      def react_on_rails_pro?
-        @react_on_rails_pro ||= gem_available?("react_on_rails_pro")
-      end
-
       def pool
-        @pool ||= if react_on_rails_pro?
+        @pool ||= if ReactOnRails::Utils.react_on_rails_pro?
                     ReactOnRailsPro::ServerRenderingPool::ProRendering
                   else
                     ReactOnRails::ServerRenderingPool::RubyEmbeddedJavaScript
@@ -25,16 +21,6 @@ module ReactOnRails
 
       def server_render_js_with_console_logging(js_code, render_options)
         pool.exec_server_render_js(js_code, render_options)
-      end
-
-      private
-
-      def gem_available?(name)
-        Gem::Specification.find_by_name(name)
-      rescue Gem::LoadError
-        false
-      rescue StandardError
-        Gem.available?(name)
       end
     end
   end