+++ lib/react_on_rails/test_helper.rb
@@ -31,7 +31,7 @@ module ReactOnRails
     # Params:
     # config - config for rspec
     # metatags - metatags to add the ensure_assets_compiled check.
-    #            Default is :js, :server_rendering
+    #            Default is :js, :server_rendering, :controller
     def self.configure_rspec_to_compile_assets(config, *metatags)
       metatags = %i[js server_rendering controller] if metatags.empty?
 
@@ -48,30 +48,31 @@ module ReactOnRails
     #                         defaults to ReactOnRails::TestHelper::WebpackAssetsStatusChecker
     # webpack_assets_compiler: provide one method: `def compile`
     #                         defaults to ReactOnRails::TestHelper::WebpackAssetsCompiler
-    # source_path and generated_assets_dir are passed into the default webpack_assets_status_checker if you
+    # source_path and generated_assets_full_path are passed into the default webpack_assets_status_checker if you
     #                        don't provide one.
     # webpack_generated_files List of files to check for up-to-date-status, defaulting to
     #                        webpack_generated_files in your configuration
     def self.ensure_assets_compiled(webpack_assets_status_checker: nil,
                                     webpack_assets_compiler: nil,
                                     source_path: nil,
-                                    generated_assets_dir: nil,
+                                    generated_assets_full_path: nil,
                                     webpack_generated_files: nil)
       ReactOnRails::WebpackerUtils.check_manifest_not_cached
       if webpack_assets_status_checker.nil?
         source_path ||= ReactOnRails::Utils.source_path
-        generated_assets_dir ||= ReactOnRails::Utils.generated_assets_dir
+        generated_assets_full_path ||= ReactOnRails::Utils.generated_assets_full_path
         webpack_generated_files ||= ReactOnRails.configuration.webpack_generated_files
 
         webpack_assets_status_checker ||=
           WebpackAssetsStatusChecker.new(source_path: source_path,
-                                         generated_assets_dir: generated_assets_dir,
+                                         generated_assets_full_path: generated_assets_full_path,
                                          webpack_generated_files: webpack_generated_files)
 
         unless @printed_once
           puts
-          puts "====> React On Rails: Checking #{webpack_assets_status_checker.generated_assets_dir} for "\
-          "outdated/missing bundles"
+          puts "====> React On Rails: Checking files in "\
+            "#{webpack_assets_status_checker.generated_assets_full_path} for "\
+            "outdated/missing bundles based on source_path #{source_path}"
           puts
           @printed_once = true
         end