+++ spec/dummy/client/app/components/ReactHelmet.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
+import HelloWorld from './HelloWorld';
 
 const EchoProps = (props) => (
   <div>
@@ -7,6 +8,7 @@ const EchoProps = (props) => (
       <title>Custom page title</title>
     </Helmet>
     Props: {JSON.stringify(props)}
+    <HelloWorld {...props}/>
   </div>
 );