+++ spec/dummy/config/initializers/react_on_rails.rb
@@ -19,7 +19,9 @@ ReactOnRails.configure do |config|
   config.node_modules_location = "client" # Pre 9.0.0 always used "client"
   config.build_production_command = "yarn run build:production"
   config.build_test_command = "yarn run build:test"
-  config.generated_assets_dir = File.join(%w[public webpack], Rails.env)
+
+  # See webpacker.yml public_output_path for replacement for generated_assets_dir
+  # config.generated_assets_dir = File.join(%w[public webpack], Rails.env)
   config.webpack_generated_files = %w[manifest.json]
   config.server_bundle_js_file = "server-bundle.js"
   config.rendering_extension = RenderingExtension