+++ spec/react_on_rails/test_helper/webpack_assets_status_checker_spec.rb
@@ -6,18 +6,20 @@ require_relative "../spec_helper"
 describe ReactOnRails::TestHelper::WebpackAssetsStatusChecker do
   describe "#stale_generated_webpack_files" do
     let(:source_path) { source_path_for(fixture_dirname) }
-    let(:generated_assets_dir) { compiled_js_dir_for(fixture_dirname) }
+    let(:generated_assets_full_path) do
+      compiled_js_dir_for(fixture_dirname)
+    end
     let(:webpack_generated_files) { %w[client-bundle.js server-bundle.js] }
-    let(:server_bundle_js_file) { File.join(generated_assets_dir, "server-bundle.js") }
-    let(:client_bundle_js_file) { File.join(generated_assets_dir, "client-bundle.js") }
+    let(:server_bundle_js_file) { File.join(generated_assets_full_path, "server-bundle.js") }
+    let(:client_bundle_js_file) { File.join(generated_assets_full_path, "client-bundle.js") }
     before do
-      allow(ReactOnRails.configuration).to receive(:generated_assets_dir)
-        .and_return(generated_assets_dir)
+      allow(ReactOnRails::WebpackerUtils).to receive(:check_manifest_not_cached).and_return(nil)
+      allow(ReactOnRails::Utils).to receive(:generated_assets_full_path).and_return(generated_assets_full_path)
     end
 
     let(:checker) do
       ReactOnRails::TestHelper::WebpackAssetsStatusChecker
-        .new(generated_assets_dir: generated_assets_dir,
+        .new(generated_assets_full_path: generated_assets_full_path,
              source_path: source_path,
              webpack_generated_files: webpack_generated_files)
     end
@@ -25,7 +27,8 @@ describe ReactOnRails::TestHelper::WebpackAssetsStatusChecker do
     context "when compiled assets exist and are up-to-date" do
       let(:fixture_dirname) { "assets_exist" }
       before do
-        touch_files_in_dir(generated_assets_dir)
+        allow(ReactOnRails::WebpackerUtils).to receive(:using_webpacker?).and_return(false)
+        touch_files_in_dir(generated_assets_full_path)
       end
 
       specify { expect(checker.stale_generated_webpack_files).to eq([]) }
@@ -39,11 +42,11 @@ describe ReactOnRails::TestHelper::WebpackAssetsStatusChecker do
         allow(ReactOnRails::WebpackerUtils).to receive(:manifest_exists?).and_return(true)
         allow(ReactOnRails::Utils).to receive(:bundle_js_file_path)
           .with("client-bundle.js")
-          .and_return(File.join(generated_assets_dir, "client-bundle-6bc530d039d96709b68d.js"))
+          .and_return(File.join(generated_assets_full_path, "client-bundle-6bc530d039d96709b68d.js"))
         allow(ReactOnRails::Utils).to receive(:bundle_js_file_path)
           .with("server-bundle.js")
-          .and_return(File.join(generated_assets_dir, "server-bundle-6bc530d039d96702268d.js"))
-        touch_files_in_dir(generated_assets_dir)
+          .and_return(File.join(generated_assets_full_path, "server-bundle-6bc530d039d96702268d.js"))
+        touch_files_in_dir(generated_assets_full_path)
       end
 
       specify { expect(checker.stale_generated_webpack_files).to eq([]) }
@@ -60,34 +63,39 @@ describe ReactOnRails::TestHelper::WebpackAssetsStatusChecker do
       specify { expect(checker.stale_generated_webpack_files).to eq(["manifest.json"]) }
     end
 
-    context "when compiled assets don't exist" do
-      let(:fixture_dirname) { "assets_no_exist" }
-
-      specify do
-        expect(checker.stale_generated_webpack_files)
-          .to eq([client_bundle_js_file, server_bundle_js_file])
-      end
-    end
-
-    context "when only server-bundle.js exists" do
-      let(:fixture_dirname) { "assets_exist_only_server_bundle" }
+    context "No Webpacker" do
       before do
-        touch_files_in_dir(generated_assets_dir)
+        allow(ReactOnRails::WebpackerUtils).to receive(:using_webpacker?).and_return(false)
       end
 
-      specify do
-        expect(checker.stale_generated_webpack_files)
-          .to eq([client_bundle_js_file])
+      context "when compiled assets don't exist" do
+        let(:fixture_dirname) { "assets_no_exist" }
+        specify do
+          expect(checker.stale_generated_webpack_files)
+            .to eq([client_bundle_js_file, server_bundle_js_file])
+        end
       end
-    end
 
-    context "when assets exist but are outdated" do
-      let(:fixture_dirname) { "assets_outdated" }
-      before { touch_files_in_dir(source_path) }
+      context "when only server-bundle.js exists" do
+        let(:fixture_dirname) { "assets_exist_only_server_bundle" }
+        before do
+          touch_files_in_dir(generated_assets_full_path)
+        end
 
-      specify do
-        expect(checker.stale_generated_webpack_files)
-          .to eq([client_bundle_js_file, server_bundle_js_file])
+        specify do
+          expect(checker.stale_generated_webpack_files)
+            .to eq([client_bundle_js_file])
+        end
+      end
+
+      context "when assets exist but are outdated" do
+        let(:fixture_dirname) { "assets_outdated" }
+        before { touch_files_in_dir(source_path) }
+
+        specify do
+          expect(checker.stale_generated_webpack_files)
+            .to eq([client_bundle_js_file, server_bundle_js_file])
+        end
       end
     end
   end